refactor(e2e): split the Docker suite from the topology it runs
A second Docker end-to-end suite is coming (4 head peers, no coils — a shape no
deployment uses, so it cannot live in the smoke test), and it would otherwise
copy ~350 lines of process orchestration.

DockerHeadSuite now holds the machinery and takes a DockerTopology; DockerSmokeTest
is the one line that binds it to what we ship. The peer counts and the compose
file are two statements of one fact that nothing reconciles — keygen-fleet writes
configs, compose separately declares services that mount them — so DockerTopology
pairs them, and rejects a coilQuorum above the coil count rather than leaving
HeadConfig to refuse it after a devnet has been built.

Note for anyone adding a subclass: CI exclusion is by FQN in build.sbt, so a new
suite must be listed there or CI will start running a minutes-long Docker test.
